Tests/LibWeb/Text/input/WebAudio/OscillatorNode.html

<script src="../include.js"></script>
<script>
    test(() => {
        const audioContext = new OfflineAudioContext(1, 5000, 44100);

        const oscillator = audioContext.createOscillator();

        // Check prototype
        let prototype = Object.getPrototypeOf(oscillator);

        while (prototype) {
            println(prototype.constructor.name);
            prototype = Object.getPrototypeOf(prototype);
        }

        // Context getter
        println(`context: '${oscillator.context}, is same as original: ${audioContext === oscillator.context}`);
    });
</script>
