Tests/LibWeb/Text/expected/WebAudio/OscillatorNode.txt

OscillatorNode
AudioScheduledSourceNode
AudioNode
EventTarget
Object
context: '[object OfflineAudioContext], is same as original: true
